feat: fix circular dep on atoms
